Updating to latest FlutterFlow output.

+++ firebase/functions/index.js
@@ -6,4 +6,14 @@ exports.onUserDeleted = functions.auth.user().onDelete(async (user) => {
   let firestore = admin.firestore();
   let userRef = firestore.doc("users/" + user.uid);
   await firestore.collection("users").doc(user.uid).delete();
+  await firestore
+    .collection("receipts")
+    .where("user_id", "==", user.uid)
+    .get()
+    .then(async (querySnapshot) => {
+      for (var doc of querySnapshot.docs) {
+        console.log(`Deleting document ${doc.id} from collection receipts`);
+        await doc.ref.delete();
+      }
+    });
 });
